FileLib Commit - Cleanup Code
Did a few changes with this commit:

- Merged fps.js with performance.js
- Removed options.json file
- In state.js, Added a control to toggle view fps
- In filelib.js removed all references to Internet Explorer
- In initial.js, cleaned up to more resemble a title screen
- Per our discussion, cleaned up a lot of comments and alerts

This code was pre-merged with the latest master code for a clean commit
and less conflicts. Keyboard functionality is working and well.
