--- new_core/json/options.json
@@ -1,8 +1,0 @@
-{
-  "up":"up",
-  "down":"down",
-  "left":"left",
-  "right":"right",
-  "select":"shift",
-  "cancel":"cancel"
-}
